refactor(objects): move ci lint to separate file
