+++ gitlab/v4/objects/__init__.py
@@ -25,6 +25,7 @@ from .badges import *
 from .boards import *
 from .branches import *
 from .broadcast_messages import *
+from .ci_lint import *
 from .clusters import *
 from .commits import *
 from .container_registry import *
