+++ tests/unit/objects/test_projects.py
@@ -82,12 +82,6 @@ pipeline_trigger_content = {
     "status": "created",
     "source": "trigger",
 }
-ci_lint_get_content = {
-    "valid": True,
-    "merged_yaml": "---\n:test_job:\n  :script: echo 1\n",
-    "errors": [],
-    "warnings": [],
-}
 
 
 @pytest.fixture
@@ -547,32 +541,6 @@ def resp_artifact():
         yield rsps
 
 
-@pytest.fixture
-def resp_get_ci_lint():
-    with responses.RequestsMock() as rsps:
-        rsps.add(
-            method=responses.GET,
-            url="http://localhost/api/v4/projects/1/ci/lint",
-            json=ci_lint_get_content,
-            content_type="application/json",
-            status=200,
-        )
-        yield rsps
-
-
-@pytest.fixture
-def resp_create_ci_lint():
-    with responses.RequestsMock() as rsps:
-        rsps.add(
-            method=responses.POST,
-            url="http://localhost/api/v4/projects/1/ci/lint",
-            json=ci_lint_get_content,
-            content_type="application/json",
-            status=200,
-        )
-        yield rsps
-
-
 def test_get_project(gl, resp_get_project):
     data = gl.projects.get(1)
     assert isinstance(data, Project)
@@ -788,17 +756,3 @@ def test_project_pull_mirror(project, resp_start_pull_mirroring_project):
 def test_project_snapshot(project, resp_snapshot_project):
     tar_file = project.snapshot()
     assert isinstance(tar_file, bytes)
-
-
-def test_project_ci_lint_get(project, resp_get_ci_lint):
-    lint_result = project.ci_lint.get()
-    assert lint_result.valid is True
-
-
-def test_project_ci_lint_create(project, resp_create_ci_lint):
-    gitlab_ci_yml = """---
-:test_job:
-  :script: echo 1
-"""
-    lint_result = project.ci_lint.create({"content": gitlab_ci_yml})
-    assert lint_result.valid is True
